refactor: simplify page content retrieval by removing countryCode par… (#2000)
* refactor: simplify page content retrieval by removing countryCode parameter

* Updated multiple page components to remove the countryCode parameter from getPageContent and getPageDetails functions.
* Adjusted RenderBuilderContent component to eliminate countryCode prop.
* Introduced new layout and client logic files for GB country code, establishing a foundation for country-specific pages.

* test: update address and SMS opt-in values in Capitol Canary advocate tests

* test: update SMS opt-in value in Capitol Canary advocate tests

* refactor: adjust builder structure again

* refactor: update generateMetadata function to use props parameter

* refactor: replace DefaultCountryCodeLayout with NotFoundLayout in NotFound component

* refactor: remove BuilderLayout

* refactor: remove BuilderLayout

* refactor: replace local user country code with default supported country code in NotFoundLayout

* chore: add TODO comment in NotFoundLayout for future countryCode implementation

* refactor: remove unused getCountryCodeFromLocalUser function from localUser utility

* refactor: remove unused imports from localUser utility

---------
