--- src/app/au/[...catchAll]/page.tsx
@@ -1,5 +1,0 @@
-import { notFound } from 'next/navigation'
-
-export default function CatchAll() {
-  notFound()
-}
