--- src/app/ca/not-found.tsx
@@ -1,5 +1,0 @@
-import { NotFoundPagesContent } from '@/components/app/notFoundPagesContent'
-
-export const dynamic = 'error'
-
-export default NotFoundPagesContent
