--- src/app/gb/not-found.tsx
@@ -1,7 +1,0 @@
-import { NotFoundPagesContent } from '@/components/app/notFoundPagesContent'
-
-export const dynamic = 'error'
-
-export default function NotFound() {
-  return <NotFoundPagesContent />
-}
