+++ src/app/not-found.tsx
@@ -1,14 +1,14 @@
 import '@/globals.css'
 
-import { DefaultCountryCodeLayout } from '@/components/app/defaultCountryCodeLayout'
+import { NotFoundLayout } from '@/components/app/notFoundLayout'
 import { NotFoundPagesContent } from '@/components/app/notFoundPagesContent'
 
 export const dynamic = 'error'
 
 export default function NotFound() {
   return (
-    <DefaultCountryCodeLayout>
+    <NotFoundLayout>
       <NotFoundPagesContent />
-    </DefaultCountryCodeLayout>
+    </NotFoundLayout>
   )
 }
